backend/requirements.txt
click==7.1.2
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
protobuf==3.13.0
# PyQt5==5.15.1
# PyQt5-sip==4.19.24
# Rtree==0.9.4
six==1.15.0
# TBB==0.1
Werkzeug==1.0.1
